feat: props로 보내줬던 컴포넌트들을 직접 페이지에서 import해서 사용
