+++ src/pages/SharePage/SharePage.js
@@ -1,14 +1,37 @@
-import { Layout } from "../../components";
+import { useState, useEffect } from "react";
+import { getFolder } from "../../api/api";
+import {
+  CardList,
+  OnlyCard,
+  FolderInfo,
+  Layout,
+  SearchBar,
+} from "../../components";
 import "./SharePage.css";
 
-const SharePage = ({ folderInfo, searchBar, cardList }) => {
+const SharePage = () => {
+  const [folderData, setFolderData] = useState([]);
+
+  const handleLoad = async () => {
+    const { folder } = await getFolder();
+    setFolderData(folder);
+  };
+
+  useEffect(() => {
+    handleLoad();
+  }, []);
+
   return (
     <Layout>
       <div className="sharePage">
-        {folderInfo}
+        <FolderInfo profile={folderData} />
         <div className="sharePage-content">
-          {searchBar}
-          {cardList}
+          <SearchBar />
+          <CardList>
+            {folderData.links?.map((item) => (
+              <OnlyCard key={item.id} items={item} {...item} />
+            ))}
+          </CardList>
         </div>
       </div>
     </Layout>
